+++ test/bigtest/tests/InvoiceEdit/InvoiceEdit-test.js
@@ -52,6 +52,12 @@ describe('Invoice edit', () => {
     expect(invoiceForm.termsInputValue).to.be.equal(TEST_VALUE_PAYMENT_TERMS);
   });
 
+  it('exchange rate fields are not presented', () => {
+    expect(invoiceForm.currentExchangeRate).to.be.false;
+    expect(invoiceForm.exchangeRate.isPresent).to.be.false;
+    expect(invoiceForm.useSetExchangeRate.isPresent).to.be.false;
+  });
+
   describe('Add data and save invoice', () => {
     const confirmation = new ConfirmationInteractor('#invoice-is-not-unique-confirmation');
 
@@ -64,6 +70,27 @@ describe('Invoice edit', () => {
     it('save button is enabled and confirmation has shown', () => {
       expect(invoiceForm.formFooter.saveButton.isDisabled).to.be.false;
       expect(confirmation.isPresent).to.be.true;
+    });
+  });
+
+  describe('Select currency other than system default currency', () => {
+    beforeEach(async function () {
+      await invoiceForm.currency.options.list(1).click();
+    });
+
+    it('current exchange rate and use set exchange rate are visible', () => {
+      expect(invoiceForm.currentExchangeRate).to.be.true;
+      expect(invoiceForm.useSetExchangeRate.isPresent).to.be.true;
+    });
+
+    describe('Check use set exchange rate', () => {
+      beforeEach(async function () {
+        await invoiceForm.useSetExchangeRate.clickAndBlur();
+      });
+
+      it('exchange rate field is visible', () => {
+        expect(invoiceForm.exchangeRate.isPresent).to.be.true;
+      });
     });
   });
 });